created CLI to simplify use
